fix seed script and refactor models hooks
