+++ tools/tsconfig.json
@@ -1,0 +1,19 @@
+{
+  "extends": "../tsconfig.base.json",
+  "compilerOptions": {
+    "module": "commonjs",
+    "outDir": "./dist",
+    "declaration": true,
+    "types": [
+      "node"
+    ]
+  },
+  "exclude": [
+    "jest.config.ts",
+    "./**/*.spec.ts",
+    "./**/*.test.ts"
+  ],
+  "include": [
+    "./**/*.ts"
+  ]
+}
